Add support for some basic reasoning 🔮
Add some limited support for the built in multi stage question
decomposer. This allows some basic reasoning and adding three
new things to the context for more complex queries.

Unfortunately there are a few bugs with the llama-index implementation
and some issues with the current prompt. This works best with llama-2-chat,
and terribly with everything else.

Due to a strange prompt inside llama-index this will ramble about
tennis, then complain there is no tennis related information in the
decompiled functions if you ask a question that is not specific enough.

We just need to feed additional context into the first query to kick
the AI down the right path.