Add base framework for interaction
Add a framework for building communications between the RE tool
and ReVa inference.

This will allow for the RE tool to make requests to ReVa for things
like renaming variables, etc.

We will later implement real time queries from ReVa to the tool
to remove the "ReVa push" requirement.

This commit adds a prototype message protocol. A server using Flask.
A example heartbeat command and an annotation for registering handlers
for commands.

To test, start the server and try:
```
curl -XPOST http://localhost:44916/project/haha-1/task \
    --json '{"message_type": "RevaHeartbeat"}'
```